Add RegLookupParse Class
No need for RegLookup.exe, thanks to @anthonyslas for finding this,
            RegistryKey localKey = RegistryKey.OpenBaseKey(RegistryHive.LocalMachine, RegistryView.Registry64);
            localKey = localKey.OpenSubKey(keyPath, false);

Which seems to bring up the intended values and doesn't run into permissions issues as other typical registry functions like Registry.GetValue() does.

This can find and list all the direct values of a key and list the sub keys as well. With the dive parameter set to true, it can also recursively go through those sub keys and and make it's own RegLookupParse object on it and then put them all listed into the SubKeys.SubkeyObj.

Co-Authored-By: Anthony Slas <anthonyslas@users.noreply.github.com>